za duzo tego syfu na razie tyle XD

+++ src/main/java/pl/vertty/arivi/utils/SpaceUtil.java
@@ -4,10 +4,10 @@
 
 package pl.vertty.arivi.utils;
 
-import java.util.Collection;
+import cn.nukkit.level.Location;
+
 import java.util.ArrayList;
 import java.util.List;
-import cn.nukkit.level.Location;
 
 public final class SpaceUtil
 {
@@ -21,7 +21,7 @@ public final class SpaceUtil
         final int maxZ = Math.max(cZ + radius, cZ - radius);
         for (int x = minX; x <= maxX; ++x) {
             for (int z = minZ; z <= maxZ; ++z) {
-                locs.add(new Location((double)x, (double)center.getFloorY(), (double)z));
+                locs.add(new Location(x, center.getFloorY(), z));
             }
         }
         locs.add(center);
@@ -36,10 +36,10 @@ public final class SpaceUtil
         final int maxX = Math.max(cX + radius, cX - radius);
         final int minZ = Math.min(cZ + radius, cZ - radius);
         final int maxZ = Math.max(cZ + radius, cZ - radius);
-        locs.add(new Location((double)minX, (double)center.getFloorY(), (double)minZ));
-        locs.add(new Location((double)maxX, (double)center.getFloorY(), (double)minZ));
-        locs.add(new Location((double)minX, (double)center.getFloorY(), (double)maxZ));
-        locs.add(new Location((double)maxX, (double)center.getFloorY(), (double)maxZ));
+        locs.add(new Location(minX, center.getFloorY(), minZ));
+        locs.add(new Location(maxX, center.getFloorY(), minZ));
+        locs.add(new Location(minX, center.getFloorY(), maxZ));
+        locs.add(new Location(maxX, center.getFloorY(), maxZ));
         return locs;
     }
     
@@ -52,7 +52,7 @@ public final class SpaceUtil
     public static List<Location> getSquare(final Location center, final int radius, final int height) {
         final List<Location> locs = getSquare(center, radius);
         for (int i = 1; i <= height; ++i) {
-            locs.addAll(getSquare(new Location((double)center.getFloorX(), (double)(center.getFloorY() + i), (double)center.getFloorZ()), radius));
+            locs.addAll(getSquare(new Location(center.getFloorX(), center.getFloorY() + i, center.getFloorZ()), radius));
         }
         return locs;
     }
@@ -60,7 +60,7 @@ public final class SpaceUtil
     public static List<Location> getCorners(final Location center, final int radius, final int height) {
         final List<Location> locs = getCorners(center, radius);
         for (int i = 1; i <= height; ++i) {
-            locs.addAll(getCorners(new Location((double)center.getFloorX(), (double)(center.getFloorY() + i), (double)center.getFloorZ()), radius));
+            locs.addAll(getCorners(new Location(center.getFloorX(), center.getFloorY() + i, center.getFloorZ()), radius));
         }
         return locs;
     }
@@ -68,7 +68,7 @@ public final class SpaceUtil
     public static List<Location> getWalls(final Location center, final int radius, final int height) {
         final List<Location> locs = getWalls(center, radius);
         for (int i = 1; i <= height; ++i) {
-            locs.addAll(getWalls(new Location((double)center.getFloorX(), (double)(center.getFloorY() + i), (double)center.getFloorZ()), radius));
+            locs.addAll(getWalls(new Location(center.getFloorX(), center.getFloorY() + i, center.getFloorZ()), radius));
         }
         return locs;
     }
@@ -83,7 +83,7 @@ public final class SpaceUtil
                 for (int y = sphere ? (cy - radius) : cy; y < (sphere ? (cy + radius) : (cy + height)); ++y) {
                     final double dist = (cx - x) * (cx - x) + (cz - z) * (cz - z) + (sphere ? ((cy - y) * (cy - y)) : 0);
                     if (dist < radius * radius && (!hollow || dist >= (radius - 1) * (radius - 1))) {
-                        final Location l = new Location((double)x, (double)(y + plusY), (double)z);
+                        final Location l = new Location(x, y + plusY, z);
                         circleblocks.add(l);
                     }
                 }
